+++ examples/escalation_policy/main.tf
@@ -1,38 +1,30 @@
-data "ilert_user" "example" {
-  email = "example@example.com"
+resource "ilert_user" "example" {
+  first_name = "example"
+  last_name  = "example"
+  email      = "example@example.com"
 }
 
-data "ilert_schedule" "example" {
-  name = "example"
+resource "ilert_schedule" "example" {
+  name     = "example"
+  timezone = "Europe/Berlin"
+  type     = "STATIC"
+  shift {
+    user  = ilert_user.example.id
+    start = "2023-09-01T08:00"
+    end   = "2023-09-02T08:00"
+  }
 }
 
 resource "ilert_escalation_policy" "example" {
   name = "example"
 
-  # escalation_rule {
-  #   escalation_timeout = 5
-  #   schedule           = data.ilert_schedule.example.id
-  # }
-
-  # escalation_rule {
-  #   escalation_timeout = 15
-  #   user               = data.ilert_user.example.id
-  # }
-
   escalation_rule {
     escalation_timeout = 15
     users {
-      id = data.ilert_user.example.id
+      id = ilert_user.example.id
     }
     schedules {
-      id = data.ilert_schedule.example.id
+      id = ilert_schedule.example.id
     }
   }
-
-  team {
-    id = 0000
-  }
-
-  # @deprecated
-  # teams = 0000
 }