+++ examples/metric_data_source/main.tf
@@ -1,15 +1,6 @@
-resource "ilert_team" "example" {
-  name = "example"
-}
-
 resource "ilert_metric_data_source" "example_prometheus" {
   name = "example"
   type = "PROMETHEUS"
-
-  team {
-    id   = ilert_team.example.id
-    name = "example"
-  }
 
   metadata {
     auth_type  = "BASIC"
@@ -19,18 +10,13 @@ resource "ilert_metric_data_source" "example_prometheus" {
   }
 }
 
-# resource "ilert_metric_data_source" "example_datadog" {
-#   name = "example"
-#   type = "DATADOG"
+resource "ilert_metric_data_source" "example_datadog" {
+  name = "example"
+  type = "DATADOG"
 
-#   team {
-#     id   = ilert_team.example.id
-#     name = "example"
-#   }
-
-#   metadata {
-#     region          = "EU1"
-#     api_key         = "your datadog api key"
-#     application_key = "your datadog application key"
-#   }
-# }
+  metadata {
+    region          = "EU1"
+    api_key         = "your datadog api key"
+    application_key = "your datadog application key"
+  }
+}