+++ examples/schedule/main.tf
@@ -1,8 +1,7 @@
 resource "ilert_user" "example" {
-  username   = "example1"
-  first_name = "John"
-  last_name  = "Doe"
-  email      = "john.doe@example.com"
+  first_name = "example"
+  last_name  = "example"
+  email      = "example@example.com"
 }
 
 # example for recurring schedule
@@ -12,7 +11,7 @@ resource "ilert_schedule" "example_recurring" {
   type     = "RECURRING"
   schedule_layer {
     name      = "layer1"
-    starts_on = "2022-08-30T00:00"
+    starts_on = "2023-08-30T00:00"
     user {
       id = ilert_user.example.id
     }
@@ -29,9 +28,6 @@ resource "ilert_schedule" "example_recurring" {
       }
     }
   }
-  team {
-    id = 0
-  }
 }
 
 # example for static schedule
@@ -41,11 +37,7 @@ resource "ilert_schedule" "example_static" {
   type     = "STATIC"
   shift {
     user  = ilert_user.example.id
-    start = "2022-09-01T08:00"
-    end   = "2022-09-02T08:00"
+    start = "2023-09-01T08:00"
+    end   = "2023-09-02T08:00"
   }
-  team {
-    id = 0
-  }
-  default_shift_duration = "PT24H"
 }