+++ examples/schedule/providers.tf
@@ -2,12 +2,12 @@ terraform {
   required_providers {
     ilert = {
       source  = "iLert/ilert"
-      version = "~> 1.6"
+      version = "~> 1.7"
     }
   }
 }
 
 provider "ilert" {
-  endpoint     = var.endpoint
-  api_token     = var.api_token
+  endpoint  = var.endpoint
+  api_token = var.api_token
 }