+++ examples/status_page/main.tf
@@ -2,9 +2,9 @@ resource "ilert_service" "example" {
   name = "example"
 }
 
-resource "ilert_status_page" "example" {
-  name       = "example"
-  subdomain  = "example.ilert.io"
+resource "ilert_status_page" "example_public" {
+  name       = "example_public"
+  subdomain  = "example-public.ilert.io"
   visibility = "PUBLIC"
 
   service {
@@ -14,13 +14,13 @@ resource "ilert_status_page" "example" {
 
 # private status page with ip whitelist enabled
 
-# resource "ilert_status_page" "example" {
-#   name         = "example"
-#   subdomain    = "example.ilert.io"
-#   visibility   = "PRIVATE"
-#   ip_whitelist = ["###.###.###.###"]
+resource "ilert_status_page" "example_private" {
+  name         = "example_private"
+  subdomain    = "example-private.ilert.io"
+  visibility   = "PRIVATE"
+  ip_whitelist = ["###.###.###.###"]
 
-#   service {
-#     id = ilert_service.example.id
-#   }
-# }
+  service {
+    id = ilert_service.example.id
+  }
+}